requirements.txt
pandas
numpy
scikit-learn
mlflow
pyyaml
python-dotenv
anthropic
jupyter
pytest
